Conflicts:
	cob_3d_mapping_pipeline/launch/mapping_demo.launch
	cob_3d_mapping_pipeline/launch/mapping_sim.launch
	cob_3d_mapping_slam/manifest.xml
	cob_3d_mapping_tools/ros/src/transform_pointcloud.cpp
	cob_3d_registration/CMakeLists.txt
	cob_3d_registration/common/include/cob_3d_registration/registration_info.h
	cob_3d_registration/common/include/registration/impl/registration_info.hpp
	cob_3d_registration/ros/launch/registration.launch
	cob_3d_registration/ros/launch/registration_demo.launch
	cob_3d_registration/ros/launch/registration_sim.launch
	cob_3d_registration/ros/src/registration_nodelet.cpp
	cob_3d_segmentation/CMakeLists.txt
	cob_3d_segmentation/common/include/cob_3d_segmentation/quad_regression/sub_structures/debug.h
	cob_3d_segmentation/common/include/cob_3d_segmentation/quad_regression/sub_structures/model.h
	cob_3d_segmentation/manifest.xml

+++ cob_3d_segmentation/common/include/cob_3d_segmentation/quad_regression/sub_structures/list.h
@@ -46,7 +46,7 @@ struct VISITED_LIST {
   }
 
   inline void add(const VALUE &v) {
-    if(size>=vals.size())
+    if((size_t)size>=vals.size())
       vals.push_back(v);
     else
       vals[size]=v;
